Fixed some serialization data types
 * This is copied from dmtcp/master commit 5c91dec (with same commit msg)

+++ src/processinfo.h
@@ -216,10 +216,10 @@ class ProcessInfo
     uint64_t _vvarEnd;
     uint64_t _endOfStack;
 
-    ptrdiff_t _clock_gettime_offset;
-    ptrdiff_t _getcpu_offset;
-    ptrdiff_t _gettimeofday_offset;
-    ptrdiff_t _time_offset;
+    uint64_t _clock_gettime_offset;
+    uint64_t _getcpu_offset;
+    uint64_t _gettimeofday_offset;
+    uint64_t _time_offset;
 };
 }
 #endif /* PROCESS_INFO */
